# laravel/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flashcards API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "description": "Get Categories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "Get resources"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "description": "Create Category",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource created"
                    }
                }
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "description": "Get Category",
                "operationId": "e92579e78391b6199e78c2a091dbea0a",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource returned"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "description": "Update Category",
                "operationId": "02aa39951e67916277e28086f85909c8",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resouce updated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "description": "Delete Category",
                "operationId": "d9ecd5c4a74c4d75339251863eedc08c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/flashcards": {
            "get": {
                "tags": [
                    "Flashcards"
                ],
                "description": "Get Flashcards",
                "operationId": "745e7a92242380cb67d3cd373765f022",
                "responses": {
                    "200": {
                        "description": "Get resources"
                    }
                }
            },
            "post": {
                "tags": [
                    "Flashcards"
                ],
                "description": "Create Flashcard",
                "operationId": "daad21d01dc5de91b535ea881fb18b52",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "translation",
                        "in": "query",
                        "description": "Translation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "theme_id",
                        "in": "query",
                        "description": "Theme ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource created"
                    }
                }
            }
        },
        "/api/flashcards/{flashcard}": {
            "get": {
                "tags": [
                    "Flashcards"
                ],
                "description": "Get Flashcard",
                "operationId": "1c30177165bc2aadc62af11696df906c",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "description": "Flashcard Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource returned"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Flashcards"
                ],
                "description": "Update Flashcard",
                "operationId": "b2ab96bb6ae33079b914b96750f36403",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "description": "Flashcard Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "translation",
                        "in": "query",
                        "description": "Translation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "theme_id",
                        "in": "query",
                        "description": "Theme ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resouce updated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Flashcards"
                ],
                "description": "Delete Flashcard",
                "operationId": "94fe4dfd3153b01a20b068e45cad6f2c",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "description": "Flashcard Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/themes": {
            "get": {
                "tags": [
                    "Themes"
                ],
                "description": "Get Themes",
                "operationId": "3af0ee5eb0876976d8f366507e93e82a",
                "responses": {
                    "200": {
                        "description": "Get resources"
                    }
                }
            },
            "post": {
                "tags": [
                    "Themes"
                ],
                "description": "Create Theme",
                "operationId": "88a51c58f980c0ba8846ebadd3a9f87a",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource created"
                    }
                }
            }
        },
        "/api/themes/{theme}": {
            "get": {
                "tags": [
                    "Themes"
                ],
                "description": "Get Theme",
                "operationId": "b26fa44f6f494cf0cbcc65d38b232cbe",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "Theme Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource returned"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Themes"
                ],
                "description": "Update Theme",
                "operationId": "2279de886d59625c857567cc236b8a50",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "Theme Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resouce updated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Themes"
                ],
                "description": "Delete Theme",
                "operationId": "0ad1a88ab068d11a34b3572bdbb2cbfa",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "Theme Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/csrf_token": {
            "get": {
                "tags": [
                    "Tokens"
                ],
                "description": "Get CSRF token",
                "operationId": "37881ba21a226ddf8572fb855a46afe1",
                "responses": {
                    "200": {
                        "description": "Resource returned"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Flashcards",
            "description": "Flashcards"
        },
        {
            "name": "Themes",
            "description": "Themes"
        },
        {
            "name": "Tokens",
            "description": "Tokens"
        }
    ]
}